requirements.txt
pesuacademy
PyPDF2
lxml
python-dotenv